Added blender script

+++ src/downsample_image.py
@@ -4,6 +4,8 @@ import os
 
 ROOT_DIR = 'data/raw/object_raw'
 
+TARGET_MAX_DIM = 512
+
 img_list = [
         f for f in glob.glob(ROOT_DIR + '/*/*.jpg')
         if not f.endswith('_mask.jpg') and not f.endswith('_ds.jpg')
@@ -15,11 +17,15 @@ for img_path in img_list:
     print(f"Processing {img_path} with shape {image.shape}")
 
     # Downsample by scale factor
-    factor = 0.25
-    downsampled = cv2.resize(image, None, fx=factor, fy=factor, interpolation=cv2.INTER_AREA)
+    h, w = image.shape[:2]
+    if max(h, w) > TARGET_MAX_DIM:
+        scale = TARGET_MAX_DIM / max(h, w)
+        new_w, new_h = int(w * scale), int(h * scale)
+        downsampled = cv2.resize(image, (new_w, new_h), interpolation=cv2.INTER_AREA)
 
-    # Save or display
-    cv2.imwrite(img_path.replace('.jpg', '_ds.jpg'), 
-                downsampled)
+        cv2.imwrite(img_path.replace('.jpg', '_ds.jpg'), downsampled)
+        os.remove(img_path)
 
-    os.remove(img_path)
+        print("Downsampled.")
+    else:
+        print("Skipped.")
